requirements.txt
streamlit
unstructured[all-docs]
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
python-dotenv